requirements.txt
pandas
google-cloud-bigquery
google-auth
pandas-gbq
db-dtypes
plotly
pybase64
requests
beautifulsoup4
Unidecode
Matplotlib
numpy